internal/socket: on Windows, use SO_RCVTIMEO to avoid blocking forever
This hack allows blocking inside WSARecvMsg indefinitely, which
then allows the runtime to wait for the socket to become readable
before trying the WSARecvMsg again.

This fixes an issue where the socket could not be Closed()d quickly,
as the Close() invocation would hang until WSARecvMsg received a
packet over the network and returned back up to RawRead() to release
the read lock on the socket.